+++ include/sdsl/suffix_tree_helper.hpp
@@ -288,6 +288,49 @@ construct_supercartesian_tree_bp_succinct_and_first_child(int_vector_buffer<t_wi
     return fc_cnt;
 }
 
+
+// Gets ISA[SA[idx]+d]
+// d = depth of the character 0 = first position
+template<class t_csa>
+typename t_csa::size_type get_char_pos(typename t_csa::size_type idx, typename t_csa::size_type d, const t_csa& csa)
+{
+    if (d == 0)
+        return idx;
+    // if we have to apply \f$\LF\f$ or \f$\Phi\f$ more
+    // than 2*d times to calc csa(csa[idx]+d), we opt to
+    // apply \f$ \Phi \f$ d times
+    if (csa.sa_sample_dens + csa.isa_sample_dens > 2*d+2) {
+        for (typename t_csa::size_type i=0; i < d; ++i)
+            idx = csa.psi[idx];
+        return idx;
+    }
+    return csa.isa[csa[idx] + d];
+}
+
+
+// has_id<X>::value is true if class X has
+// implement method id
+// Adapted solution from jrok's proposal:
+// http://stackoverflow.com/questions/87372/check-if-a-class-has-a-member-function-of-a-given-signature
+template<typename t_wt>
+struct has_id {
+    template<typename T>
+    static constexpr auto check(T*)
+    -> typename
+    std::is_same<
+    decltype(std::declval<T>().id(
+                 std::declval<typename T::node_type&>()
+             )),
+             typename T::size_type>::type {return std::true_type();}
+             template<typename>
+    static constexpr std::false_type check(...) {return std::false_type();}
+    typedef decltype(check<t_wt>(nullptr)) type;
+    static constexpr bool value = type::value;
+};
+
+
+
+
 }
 
 #endif
