+++ include/sdsl/suffix_trees.hpp
@@ -21,64 +21,6 @@
 #ifndef INCLUDED_SDSL_SUFFIX_TREES
 #define INCLUDED_SDSL_SUFFIX_TREES
 
-#include "sdsl_concepts.hpp"
-#include "suffix_arrays.hpp"
-#include "suffix_tree_algorithm.hpp"
-#include "construct.hpp"
-#include "util.hpp"
-#include <iostream>
-#include <cmath>
-#include <set>
-
-using std::cout;
-using std::endl;
-
-namespace sdsl
-{
-
-// Gets ISA[SA[idx]+d]
-// d = depth of the character 0 = first position
-template<class t_csa>
-typename t_csa::size_type get_char_pos(typename t_csa::size_type idx, typename t_csa::size_type d, const t_csa& csa)
-{
-    if (d == 0)
-        return idx;
-    // if we have to apply \f$\LF\f$ or \f$\Phi\f$ more
-    // than 2*d times to calc csa(csa[idx]+d), we opt to
-    // apply \f$ \Phi \f$ d times
-    if (csa.sa_sample_dens + csa.isa_sample_dens > 2*d+2) {
-        for (typename t_csa::size_type i=0; i < d; ++i)
-            idx = csa.psi[idx];
-        return idx;
-    }
-    return csa.isa[csa[idx] + d];
-}
-
-
-// has_id<X>::value is true if class X has
-// implement method id
-// Adapted solution from jrok's proposal:
-// http://stackoverflow.com/questions/87372/check-if-a-class-has-a-member-function-of-a-given-signature
-template<typename t_wt>
-struct has_id {
-    template<typename T>
-    static constexpr auto check(T*)
-    -> typename
-    std::is_same<
-    decltype(std::declval<T>().id(
-                 std::declval<typename T::node_type&>()
-             )),
-             typename T::size_type>::type {return std::true_type();}
-             template<typename>
-    static constexpr std::false_type check(...) {return std::false_type();}
-    typedef decltype(check<t_wt>(nullptr)) type;
-    static constexpr bool value = type::value;
-};
-
-
-
-}
-
 /** \defgroup cst Compressed Suffix Trees (CST)
  *   This group contains data structures for compressed suffix trees. The following methods are supported:
  *    - root()
@@ -91,15 +33,10 @@ struct has_id {
  *    - ..
  */
 
-#include "suffix_tree_helper.hpp"
 #include "cst_sct3.hpp"
 #include "cst_sada.hpp"
 #include "cst_fully.hpp"
 
-#include "csa_bitcompressed.hpp"
-#include "int_vector.hpp"
 
-#include <iostream>
-#include <string>
 
 #endif
